+++ src/mut_bump_vec_rev.rs
@@ -1,6 +1,6 @@
 use crate::{
     destructure::destructure,
-    error_behavior_generic_methods_allocation_failure, owned_slice,
+    error_behavior_generic_methods_allocation_failure, min_non_zero_cap, owned_slice,
     polyfill::{self, nonnull, pointer},
     BaseAllocator, BumpBox, BumpScope, ErrorBehavior, GuaranteedAllocatedStats, MinimumAlignment, NoDrop, SetLenOnDrop,
     SizedTypeProperties, Stats, SupportedMinimumAlignment,
@@ -1335,6 +1335,8 @@ where
         }
 
         let new_cap = self.capacity().checked_mul(2).unwrap_or(required_cap).max(required_cap);
+        let new_cap = new_cap.max(min_non_zero_cap(T::SIZE));
+
         unsafe { self.generic_grow_to(new_cap) }
     }