+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- **added:** minimum capacity when growing vectors and strings (just like `Vec`)
+
 ## 0.10.6 (2024-10-26)
 - **fixed:** potential UB when using `BumpVec::splice`
 - **added:** default generic allocator parameter for `bump_vec::Splice`